Add exceptions and commands & tasks packages
