+++ src/main/java/commands/Command.java
@@ -1,3 +1,6 @@
+package commands;
+import tasks.TaskManager;
+
 public abstract class Command {
     public abstract String execute(TaskManager taskManager) throws DukeException;
 }
