+++ src/main/java/commands/DeleteCommand.java
@@ -1,3 +1,7 @@
+package commands;
+import tasks.Task;
+import tasks.TaskManager;
+
 public class DeleteCommand extends Command {
     private final int taskNumber;
     public static final String KEYWORD = "delete";
