+++ src/main/java/commands/DukeException.java
@@ -1,3 +1,5 @@
+package commands;
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
