+++ src/main/java/commands/InvalidCommand.java
@@ -1,3 +1,6 @@
+package commands;
+import tasks.TaskManager;
+
 public class InvalidCommand extends Command {
     public String execute(TaskManager taskManager) throws DukeException {
         throw new DukeException("OOPS!!! I'm sorry, but I don't know what that means :-(");
