+++ src/main/java/tasks/Task.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class Task {
     private final String taskName;
     private boolean isDone;
