+++ src/main/java/tasks/TaskManager.java
@@ -1,3 +1,7 @@
+package tasks;
+
+import tasks.Task;
+
 import java.util.ArrayList;
 
 public class TaskManager {
